refactor: 移除不必要的类型定义和接口，优化导入结构
