+++ packages/specification/src/common/interaction.ts
@@ -4,7 +4,6 @@
  */
 
 import type { CommonElement } from './elements';
-import type { AnimationProperties } from '../core/interfaces';
 
 /**
  * 交互事件类型
@@ -445,212 +444,4 @@ export enum HitAreaType {
    * 像素完美
    */
   PixelPerfect = 'pixel-perfect',
-}
-
-/**
- * 悬停效果
- */
-export interface HoverEffect {
-  /**
-   * 是否启用
-   */
-  enabled: boolean;
-  /**
-   * 悬停延迟（毫秒）
-   */
-  delay: number;
-  /**
-   * 高亮颜色
-   */
-  highlightColor?: [number, number, number, number];
-  /**
-   * 缩放因子
-   */
-  scaleFactor?: number;
-  /**
-   * 透明度变化
-   */
-  opacityChange?: number;
-  /**
-   * 动画配置
-   */
-  animation?: AnimationProperties;
-  /**
-   * 光标样式
-   */
-  cursor?: string;
-  /**
-   * 工具提示
-   */
-  tooltip?: string;
-}
-
-/**
- * 点击效果
- */
-export interface ClickEffect {
-  /**
-   * 是否启用
-   */
-  enabled: boolean;
-  /**
-   * 反馈类型
-   */
-  feedbackType: ClickFeedbackType;
-  /**
-   * 动画配置
-   */
-  animation?: AnimationProperties;
-  /**
-   * 音效
-   */
-  sound?: string;
-  /**
-   * 震动
-   */
-  vibration?: VibrationPattern;
-  /**
-   * 视觉效果
-   */
-  visualEffect?: VisualEffect;
-}
-
-/**
- * 点击反馈类型
- */
-export enum ClickFeedbackType {
-  /**
-   * 无反馈
-   */
-  None = 'none',
-  /**
-   * 缩放
-   */
-  Scale = 'scale',
-  /**
-   * 闪烁
-   */
-  Flash = 'flash',
-  /**
-   * 波纹
-   */
-  Ripple = 'ripple',
-  /**
-   * 弹跳
-   */
-  Bounce = 'bounce',
-  /**
-   * 旋转
-   */
-  Rotate = 'rotate',
-}
-
-/**
- * 震动模式
- */
-export interface VibrationPattern {
-  /**
-   * 震动持续时间（毫秒）
-   */
-  duration: number;
-  /**
-   * 震动强度 (0-1)
-   */
-  intensity: number;
-  /**
-   * 震动模式
-   */
-  pattern?: number[];
-}
-
-/**
- * 视觉效果
- */
-export interface VisualEffect {
-  /**
-   * 效果类型
-   */
-  type: VisualEffectType;
-  /**
-   * 持续时间（毫秒）
-   */
-  duration: number;
-  /**
-   * 效果参数
-   */
-  parameters?: Record<string, any>;
-}
-
-/**
- * 视觉效果类型
- */
-export enum VisualEffectType {
-  /**
-   * 粒子效果
-   */
-  Particle = 'particle',
-  /**
-   * 光晕效果
-   */
-  Glow = 'glow',
-  /**
-   * 闪光效果
-   */
-  Flash = 'flash',
-  /**
-   * 波纹效果
-   */
-  Ripple = 'ripple',
-  /**
-   * 爆炸效果
-   */
-  Explosion = 'explosion',
-}
-
-/**
- * 选择效果
- */
-export interface SelectionEffect {
-  /**
-   * 是否启用
-   */
-  enabled: boolean;
-  /**
-   * 是否多选
-   */
-  multiSelect: boolean;
-  /**
-   * 选择颜色
-   */
-  selectionColor?: [number, number, number, number];
-  /**
-   * 选择边框
-   */
-  selectionBorder?: SelectionBorder;
-  /**
-   * 选择动画
-   */
-  animation?: AnimationProperties;
-}
-
-/**
- * 选择边框
- */
-export interface SelectionBorder {
-  /**
-   * 边框宽度
-   */
-  width: number;
-  /**
-   * 边框颜色
-   */
-  color: [number, number, number, number];
-  /**
-   * 边框样式
-   */
-  style: 'solid' | 'dashed' | 'dotted';
-  /**
-   * 边框圆角
-   */
-  radius?: number;
 }
