+++ packages/specification/src/common/material.ts
@@ -3,50 +3,8 @@
  * 定义所有系统共通的材质相关类型
  */
 
-import type { Color } from '../core/interfaces';
+import type { MaterialProperties } from '../core';
 import type { BlendMode } from '../core/enums';
-
-/**
- * 材质类型
- */
-export enum MaterialType {
-  /**
-   * 标准材质
-   */
-  Standard = 'standard',
-  /**
-   * 无光照材质
-   */
-  Unlit = 'unlit',
-  /**
-   * 物理材质
-   */
-  Physical = 'physical',
-  /**
-   * 卡通材质
-   */
-  Toon = 'toon',
-  /**
-   * 精灵材质
-   */
-  Sprite = 'sprite',
-  /**
-   * UI材质
-   */
-  UI = 'ui',
-  /**
-   * 粒子材质
-   */
-  Particle = 'particle',
-  /**
-   * 天空盒材质
-   */
-  Skybox = 'skybox',
-  /**
-   * 自定义材质
-   */
-  Custom = 'custom',
-}
 
 /**
  * 纹理类型
@@ -144,116 +102,6 @@ export enum TextureWrap {
    * 边界夹紧
    */
   ClampToBorder = 'clamp-to-border',
-}
-
-/**
- * 材质属性
- */
-export interface MaterialProperties {
-  /**
-   * 材质名称
-   */
-  name: string;
-  /**
-   * 材质类型
-   */
-  type: MaterialType;
-  /**
-   * 基础颜色
-   */
-  baseColor: Color;
-  /**
-   * 透明度
-   */
-  opacity: number;
-  /**
-   * 金属度
-   */
-  metallic?: number;
-  /**
-   * 粗糙度
-   */
-  roughness?: number;
-  /**
-   * 自发光颜色
-   */
-  emissiveColor?: Color;
-  /**
-   * 自发光强度
-   */
-  emissiveIntensity?: number;
-  /**
-   * 法线强度
-   */
-  normalScale?: number;
-  /**
-   * 遮挡强度
-   */
-  occlusionStrength?: number;
-  /**
-   * 折射率
-   */
-  ior?: number;
-  /**
-   * 透射
-   */
-  transmission?: number;
-  /**
-   * 厚度
-   */
-  thickness?: number;
-  /**
-   * 衰减颜色
-   */
-  attenuationColor?: Color;
-  /**
-   * 衰减距离
-   */
-  attenuationDistance?: number;
-  /**
-   * 各向异性
-   */
-  anisotropy?: number;
-  /**
-   * 各向异性旋转
-   */
-  anisotropyRotation?: number;
-  /**
-   * 清漆
-   */
-  clearcoat?: number;
-  /**
-   * 清漆粗糙度
-   */
-  clearcoatRoughness?: number;
-  /**
-   * 清漆法线
-   */
-  clearcoatNormal?: number;
-  /**
-   * 光泽
-   */
-  sheen?: number;
-  /**
-   * 光泽颜色
-   */
-  sheenColor?: Color;
-  /**
-   * 光泽粗糙度
-   */
-  sheenRoughness?: number;
-  /**
-   * 次表面散射
-   */
-  subsurface?: number;
-  /**
-   * 次表面颜色
-   */
-  subsurfaceColor?: Color;
-  /**
-   * 次表面半径
-   */
-  subsurfaceRadius?: [number, number, number];
 }
 
 /**
