+++ packages/specification/src/common/transform.ts
@@ -3,147 +3,8 @@
  * 定义所有系统共通的变换相关类型
  */
 
+import type { Matrix3x3, Matrix4x4, Quaternion, RotationOrder, TransformSpace, Vector2, Vector3 } from '../core';
 import type { UsdValue } from '../core/usd';
-
-/**
- * 2D向量
- */
-export interface Vector2 {
-  x: number;
-  y: number;
-}
-
-/**
- * 3D向量
- */
-export interface Vector3 {
-  x: number;
-  y: number;
-  z: number;
-}
-
-/**
- * 4D向量
- */
-export interface Vector4 {
-  x: number;
-  y: number;
-  z: number;
-  w: number;
-}
-
-/**
- * 四元数
- */
-export interface Quaternion {
-  x: number;
-  y: number;
-  z: number;
-  w: number;
-}
-
-/**
- * 2D矩阵
- */
-export interface Matrix2x2 {
-  m00: number;
-  m01: number;
-  m10: number;
-  m11: number;
-}
-
-/**
- * 3D矩阵
- */
-export interface Matrix3x3 {
-  m00: number;
-  m01: number;
-  m02: number;
-  m10: number;
-  m11: number;
-  m12: number;
-  m20: number;
-  m21: number;
-  m22: number;
-}
-
-/**
- * 4D矩阵
- */
-export interface Matrix4x4 {
-  m00: number;
-  m01: number;
-  m02: number;
-  m03: number;
-  m10: number;
-  m11: number;
-  m12: number;
-  m13: number;
-  m20: number;
-  m21: number;
-  m22: number;
-  m23: number;
-  m30: number;
-  m31: number;
-  m32: number;
-  m33: number;
-}
-
-/**
- * 变换空间
- */
-export enum TransformSpace {
-  /**
-   * 世界空间
-   */
-  World = 'world',
-  /**
-   * 本地空间
-   */
-  Local = 'local',
-  /**
-   * 父级空间
-   */
-  Parent = 'parent',
-  /**
-   * 屏幕空间
-   */
-  Screen = 'screen',
-  /**
-   * 视图空间
-   */
-  View = 'view',
-}
-
-/**
- * 旋转顺序
- */
-export enum RotationOrder {
-  /**
-   * XYZ顺序
-   */
-  XYZ = 'xyz',
-  /**
-   * XZY顺序
-   */
-  XZY = 'xzy',
-  /**
-   * YXZ顺序
-   */
-  YXZ = 'yxz',
-  /**
-   * YZX顺序
-   */
-  YZX = 'yzx',
-  /**
-   * ZXY顺序
-   */
-  ZXY = 'zxy',
-  /**
-   * ZYX顺序
-   */
-  ZYX = 'zyx',
-}
 
 /**
  * 通用2D变换
@@ -309,28 +170,6 @@ export interface TransformConstraint {
    * 约束参数
    */
   parameters?: Record<string, any>;
-}
-
-/**
- * 边界框
- */
-export interface BoundingBox {
-  /**
-   * 最小点
-   */
-  min: Vector3;
-  /**
-   * 最大点
-   */
-  max: Vector3;
-  /**
-   * 中心点
-   */
-  center?: Vector3;
-  /**
-   * 尺寸
-   */
-  size?: Vector3;
 }
 
 /**
