+++ packages/specification/src/index.ts
@@ -2,10 +2,10 @@
 export * from './core';
 
 // 其他模块使用命名空间导出避免冲突
-export * as Common from './common';
-export * as Animation from './animation';
-export * as Design from './design';
-export * as Media from './media';
-export * as Rendering from './rendering';
-export * as Workflow from './workflow';
-export * as Package from './package';
+export * from './common';
+// export * as Animation from './animation';
+// export * as Design from './design';
+// export * as Media from './media';
+// export * as Rendering from './rendering';
+// export * as Workflow from './workflow';
+// export * as Package from './package';
